feat: Added capture groups to text based alerts
You can use `{}` to surround parts of a message you want to capture and then use `$n` in a notification where n is the group number.
e.g. `Your {*} is ready to harvest in {*}` for the chat alert and in the Text to Speech you could put `The $1 is ready in $2`

Fixes #9